refactor: move raft layer folder
