+++ cmd/crond.go
@@ -11,7 +11,7 @@ import (
 	"syscall"
 	"time"
 
-	"github.com/KevinWu0904/crond/internal/raft"
+	"github.com/KevinWu0904/crond/internal/crond/raft"
 
 	"github.com/KevinWu0904/crond/internal/config"
 	"github.com/KevinWu0904/crond/internal/crond"
